removing/remplacing nt2/table.hpp include

+++ modules/core/linalg/include/nt2/toolbox/linalg/functions/details/symeig.hpp
@@ -150,7 +150,7 @@ namespace nt2 { namespace details
     // /////////////////////////////////////////////////////////////////////////////
     size_t     rank(base_t epsi = nt2::Eps<base_t>())
     {
-      int32_t r = 0;
+      size_t r = 0;
       base_t thresh = n_*epsi*nt2::max(nt2::abs(w_(_)));
       for(int i=1; i <= n_; ++i)
         {
